pi/requirements.txt
customtkinter
pyserial
opencv-python
pillow
numpy
pandas
matplotlib
scipy
tk